Adding in the initial wireframe for motor control
This works on one motor in the z axis which was the most obvious to
observe on the camera at FXI. This has a number of flaws including the
hardwiring of the motor PV names, a growing main Python file that needs
refactoring.

It does show how to use the Ophyd object to set the motor position, and
turning subscribe callbacks from a separate thread into Python signals
that can update motor positions without a polling loop. Next steps will
refactor the code to better capture the different elements, and make the
motor PV names configurable along with exploring interface options for
the x, y, z and rotation PVs.