More widgets for moving the axes and a plan!
Added more widgets to get the interaction with moves, step sizes etc
right. This needs some level of de-bouncing and waiting for previous
requests to finish up.

I also added the very simplest plan, moving the rotation back to 0. This
was all tested at FXI today when there was no beam to look at
interactivity of the system including the RE-0MQ piece.